pyproject.toml
[build-system]
requires = [
    "hatchling>=1.8.1",
]
build-backend = "hatchling.build"

[project]
name = "parsedmarc"
dynamic = [
    "version",
]
description = "A Python package and CLI for parsing aggregate and forensic DMARC reports"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    { name = "Sean Whalen", email = "whalenster@gmail.com" },
]
keywords = [
    "DMARC",
    "parser",
    "reporting",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

requires-python = ">=3.7"

dependencies = [
    "azure-identity>=1.8.0",
    "azure-monitor-ingestion>=1.0.0",
    "boto3>=1.16.63",
    "bs4",
    "dateparser>=1.1.1",
    "dnspython>=2.0.0",
    "elasticsearch-dsl==7.4.0",
    "elasticsearch<7.14.0",
    "expiringdict>=1.1.4",
    "geoip2>=3.0.0",
    "google-api-core>=2.4.0",
    "google-api-python-client>=2.35.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.4.6",
    "google-auth>=2.3.3",
    "imapclient>=2.1.0",
    "kafka-python>=1.4.4",
    "lxml>=4.4.0",
    "mailsuite>=1.6.1",
    "msgraph-core>=0.2.2",
    "publicsuffixlist>=0.10.0",
    "requests>=2.22.0",
    "tqdm>=4.31.1",
    "urllib3>=1.25.7",
    "xmltodict>=0.12.0",
]

[project.optional-dependencies]
dev = [
    ## Type Stubs
    "google-api-python-client-stubs",
    "google-auth-stubs",
    "lxml-stubs",
    "types-boto3",
    "types-python-dateutil",
    "types-requests",
    "types-tqdm",
    "types-xmltodict",
    ## Formatting / Linting
    "validate-pyproject[all]",
    "pyupgrade",
    "black",
    "isort",
    #"pylint",
    "mypy",
    "flake8",
    ## Testing
    "codecov",
]

docs = [
    "alabaster>=0.7.12",
    "Babel>=2.7.0",
    "docutils>=0.14,<0.18",
    "myst-parser[linkify]>=0.18.0",
    "pygments>=2.11.1",
    "sphinx>=1.0.5",
    "sphinx_rtd_theme>=0.4.3",
]


[project.scripts]
parsedmarc = "parsedmarc.cli:_main"

[project.urls]
Homepage = "https://domainaware.github.io/parsedmarc"

[tool.hatch.version]
path = "parsedmarc/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/parsedmarc",
]

[tool.isort]
profile = "black"
force_sort_within_sections = true
src_paths = ["parsedmarc", "tests.py"]

# Section Headings
import_heading_future = "Future"
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Installed"
import_heading_firstparty = "Package"
import_heading_localfolder = "Local"

[tool.black]
line-length = 100
target-version = ["py37", "py38", "py39", "py310", "py311", "py312"]
